+++ src/assets/locales/hi.js
