+++ src/components/UI/PlioThumbnail.vue
