+++ src/components/UI/tutorial/ProceedButtonPointer.vue
