+++ src/components/UI/tutorial/StartButtonPointer.vue
