+++ src/components/UI/tutorial/mcqOptionsPointer.vue
