+++ src/router/index.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHashHistory } from "vue-router";
-import Home from "@/views/Home.vue";
-import Player from "@/views/Player.vue";
-import Login from "@/views/Login";
+import Home from "@/pages/Home.vue";
+import Player from "@/pages/Player.vue";
+import PhoneSignIn from "@/pages/PhoneSignIn";
 
 const routes = [
   {
@@ -27,8 +27,8 @@ const routes = [
     // type: the type of component invoking this path (optional)
     // id: the unique ID for the component invoking this path (optional)
     path: "/login/:id?/:type?",
-    name: "Login",
-    component: Login,
+    name: "PhoneSignIn",
+    component: PhoneSignIn,
     // passing props to route components
     // https://router.vuejs.org/guide/essentials/passing-props.html#passing-props-to-route-components
     props: true,
@@ -38,13 +38,13 @@ const routes = [
     name: "ABTesting",
     // lazy loading of routes
     // https://router.vuejs.org/guide/advanced/lazy-loading.html#grouping-components-in-the-same-chunk
-    component: () => import("@/components/FunctionalComponents/ABTesting"),
+    component: () => import("@/utils/ABTesting"),
     props: true,
   },
   {
     path: "/404-not-found",
     name: "404",
-    component: () => import("../views/Error"),
+    component: () => import("@/pages/Error"),
     props: { type: "404" },
   },
   {
